modularize and reuse 2D helper functions
