+++ src/bench_func.py
@@ -14,7 +14,7 @@ from joblib import Parallel, delayed
 import copy
 
 import src.utility.convergence as convergence
-from src.benchmark_models import settings_2Dchromatography
+import src.benchmark_models.helper_setup_2Dmodels as helper2D
 
 
 # %% Import packages and define helper functions
@@ -312,7 +312,7 @@ def create_object_from_config(
                     n_units = (rad_method + 1 ) * rad_cells + 1 + nOutlets
                 config_data['input']['model'].nunits = n_units
                 
-                connections, rad_coords = settings_2Dchromatography.generate_connections_matrix(
+                connections, rad_coords = helper2D.generate_connections_matrix(
                     rad_method=rad_method, rad_cells=rad_cells,
                     velocity=config_data['input']['model']['unit_' +
                                                            tmpID].VELOCITY,
