+++ src/benchmark_models/setting_Col2D_SMA_4comp_LWE_benchmark1.py
@@ -2,7 +2,7 @@ import numpy as np
 from addict import Dict
 import copy
 
-import src.benchmark_models.helper_connections_2Dmodels as helper
+import src.benchmark_models.helper_setup_2Dmodels as helper
 
 
 def get_model(
